requirements.txt
Django==5.0.6
djangorestframework==3.15.1
easyocr==1.7.1
numpy==1.26.2
Pillow==9.4.0
gunicorn
deepface==0.0.91
tf-keras
ultralytics==8.2.26
opencv-python==4.9.0.80
matplotlib==3.9.0
Pillow==9.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
mediapipe==0.10.14
django-cors-headers
spacy==3.7.5
python-bidi==0.4.2
onnxruntime==1.18.1
python-bidi==0.4.2
